pyproject.toml
[project]
name = "sentinela"
version = "3.1.1"
description = "Experimental open-source EDR with local AI analysis (Ollama)"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Miguel F Araujo", email = "nigelkps1@gmail.com" },
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "requests",
    "psutil",
    "python-nmap",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
sentinela = "app.agente:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
